Refactor Brand arquitecture

+++ ApiComponents.Application/Features/Brands/Commands/DeleteBrandCommand.cs
@@ -3,13 +3,17 @@ using MediatR;
 
 namespace ApiComponents.Application.Features.Brands.Commands;
 
-public record DeleteBrandCommand(int Id) : IRequest<bool>;
+public record DeleteBrandCommand(int Id) : IRequest<Unit>;
 
-public class DeleteBrandCommandHandler(IBrandRepository brandRepository) : IRequestHandler<DeleteBrandCommand, bool>
+public class DeleteBrandCommandHandler(IBrandRepository brandRepository) : IRequestHandler<DeleteBrandCommand, Unit>
 {
-    public async Task<bool> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
+    public async Task<Unit> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
     {
+        var brand = await brandRepository.GetBrandByIdAsync(request.Id, cancellationToken);
+        if (brand == null)
+            throw new ApplicationException($"La marca con ID {request.Id} no existe.");
+
         await brandRepository.DeleteBrandAsync(request.Id, cancellationToken);
-        return true;
+        return Unit.Value;
     }
 }
